requirements.txt
asgiref==3.4.1
astroid==2.5.6
certifi==2020.12.5
configparser==5.0.2
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
minio==7.0.3
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
pylint==2.8.2
python-dateutil==2.8.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
